requirements.txt
psycopg2-binary>=2.9.0
google-genai>=0.2.0

